refactor(inflight): split save-store identity gate
