+++ ARCHITECTURE.md
@@ -465,6 +465,8 @@ src/
 │   │   │   ├── clear_store/
 │   │   │   │   ├── abandon.rs
 │   │   │   │   └── mod.rs
+│   │   │   ├── save_store/
+│   │   │   │   └── identity_gate.rs
 │   │   │   ├── anchor_repost.rs
 │   │   │   ├── budget.rs
 │   │   │   ├── finalizer_identity.rs
